Solving problem with classes bad placed
