+++ library/src/main/java/com/mycompany/library/Menu.form
@@ -73,7 +73,7 @@
               <Font name="Liberation Serif" size="24" style="0"/>
             </Property>
             <Property name="icon" type="javax.swing.Icon" editor="org.netbeans.modules.form.editors2.IconEditor">
-              <Image iconType="3" name="/Images/exit.png"/>
+              <Image iconType="3" name="/com/mycompany/library/images/exit.png"/>
             </Property>
             <Property name="text" type="java.lang.String" value="Salir"/>
           </Properties>
@@ -82,7 +82,7 @@
           </Events>
           <Constraints>
             <Constraint layoutClass="org.netbeans.modules.form.compat2.layouts.DesignAbsoluteLayout" value="org.netbeans.modules.form.compat2.layouts.DesignAbsoluteLayout$AbsoluteConstraintsDescription">
-              <AbsoluteConstraints x="560" y="730" width="210" height="-1"/>
+              <AbsoluteConstraints x="570" y="730" width="210" height="-1"/>
             </Constraint>
           </Constraints>
         </Component>
@@ -160,7 +160,7 @@
                   <Font name="Liberation Serif" size="24" style="0"/>
                 </Property>
                 <Property name="icon" type="javax.swing.Icon" editor="org.netbeans.modules.form.editors2.IconEditor">
-                  <Image iconType="3" name="/Images/book.png"/>
+                  <Image iconType="3" name="/com/mycompany/library/images/book.png"/>
                 </Property>
                 <Property name="text" type="java.lang.String" value="Libros"/>
               </Properties>
@@ -174,7 +174,7 @@
                   <Font name="Liberation Serif" size="24" style="0"/>
                 </Property>
                 <Property name="icon" type="javax.swing.Icon" editor="org.netbeans.modules.form.editors2.IconEditor">
-                  <Image iconType="3" name="/Images/students.png"/>
+                  <Image iconType="3" name="/com/mycompany/library/images/students.png"/>
                 </Property>
                 <Property name="text" type="java.lang.String" value="Estudiantes"/>
               </Properties>
@@ -188,7 +188,7 @@
                   <Font name="Liberation Serif" size="24" style="0"/>
                 </Property>
                 <Property name="icon" type="javax.swing.Icon" editor="org.netbeans.modules.form.editors2.IconEditor">
-                  <Image iconType="3" name="/Images/loan.png"/>
+                  <Image iconType="3" name="/com/mycompany/library/images/loan.png"/>
                 </Property>
                 <Property name="text" type="java.lang.String" value="Pr&#xe9;stamos"/>
               </Properties>
@@ -199,7 +199,7 @@
                   <Font name="Liberation Serif" size="24" style="0"/>
                 </Property>
                 <Property name="icon" type="javax.swing.Icon" editor="org.netbeans.modules.form.editors2.IconEditor">
-                  <Image iconType="3" name="/Images/report.png"/>
+                  <Image iconType="3" name="/com/mycompany/library/images/report.png"/>
                 </Property>
                 <Property name="text" type="java.lang.String" value="Reportes"/>
               </Properties>
@@ -209,6 +209,25 @@
             </Component>
           </SubComponents>
         </Container>
+        <Component class="javax.swing.JButton" name="btnExit1">
+          <Properties>
+            <Property name="font" type="java.awt.Font" editor="org.netbeans.beaninfo.editors.FontEditor">
+              <Font name="Liberation Serif" size="24" style="0"/>
+            </Property>
+            <Property name="icon" type="javax.swing.Icon" editor="org.netbeans.modules.form.editors2.IconEditor">
+              <Image iconType="3" name="/com/mycompany/library/images/importData.png"/>
+            </Property>
+            <Property name="text" type="java.lang.String" value="Importar datos"/>
+          </Properties>
+          <Events>
+            <EventHandler event="actionPerformed" listener="java.awt.event.ActionListener" parameters="java.awt.event.ActionEvent" handler="btnExit1ActionPerformed"/>
+          </Events>
+          <Constraints>
+            <Constraint layoutClass="org.netbeans.modules.form.compat2.layouts.DesignAbsoluteLayout" value="org.netbeans.modules.form.compat2.layouts.DesignAbsoluteLayout$AbsoluteConstraintsDescription">
+              <AbsoluteConstraints x="280" y="630" width="230" height="-1"/>
+            </Constraint>
+          </Constraints>
+        </Component>
       </SubComponents>
     </Container>
   </SubComponents>
