+++ library/src/main/java/com/mycompany/library/images/addBook.png
