+++ library/src/main/java/com/mycompany/library/images/book.png
