+++ library/src/main/java/com/mycompany/library/images/exit.png
