+++ library/src/main/java/com/mycompany/library/images/loan.png
