+++ library/src/main/java/com/mycompany/library/images/search.png
