+++ library/src/main/java/com/mycompany/library/students/StudentsForm.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
  */
-package Books;
+package com.mycompany.library.students;
 
 import com.mycompany.library.Menu;
 
@@ -10,19 +10,19 @@ import com.mycompany.library.Menu;
  *
  * @author carlos
  */
-public class booksForm extends javax.swing.JFrame {
+public class StudentsForm extends javax.swing.JFrame {
 
     /**
      * Creates new form StudentsForm
      */
-    public booksForm(Menu menu) {
+    public StudentsForm(Menu menu) {
         initComponents();
         menu = this.menu;
     }
     
     Menu menu = new Menu();
 
-    private booksForm() {
+    private StudentsForm() {
         throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
     }
 
@@ -41,8 +41,8 @@ public class booksForm extends javax.swing.JFrame {
         jPanel2 = new javax.swing.JPanel();
         btnAdd = new javax.swing.JButton();
         lblNewStudent1 = new javax.swing.JLabel();
-        txtTitle = new javax.swing.JTextField();
-        txtAutor = new javax.swing.JTextField();
+        txtName = new javax.swing.JTextField();
+        txtCarnet = new javax.swing.JTextField();
         lblNewStudent2 = new javax.swing.JLabel();
         lblNewStudent3 = new javax.swing.JLabel();
         txtdd = new javax.swing.JTextField();
@@ -55,28 +55,15 @@ public class booksForm extends javax.swing.JFrame {
         lblNewStudent9 = new javax.swing.JLabel();
         lblNewStudent10 = new javax.swing.JLabel();
         lblNewStudent11 = new javax.swing.JLabel();
-        lblNewStudent12 = new javax.swing.JLabel();
-        btnImport = new javax.swing.JButton();
-        txtCodBook = new javax.swing.JTextField();
-        lblNewStudent13 = new javax.swing.JLabel();
-        txtNoCopies = new javax.swing.JTextField();
-        lblNewStudent14 = new javax.swing.JLabel();
-        txtEditorial = new javax.swing.JTextField();
+        jcbCodCarrera = new javax.swing.JComboBox<>();
         jPanel3 = new javax.swing.JPanel();
         lblNewStudent = new javax.swing.JLabel();
         lblNewStudent5 = new javax.swing.JLabel();
         jcbFilter = new javax.swing.JComboBox<>();
         txtFilter = new javax.swing.JTextField();
         tblStudents = new javax.swing.JScrollPane();
-        tblBooks = new javax.swing.JTable();
+        jTable2 = new javax.swing.JTable();
         btnSearch = new javax.swing.JButton();
-        jPanel4 = new javax.swing.JPanel();
-        lblNewStudent15 = new javax.swing.JLabel();
-        lblNewStudent16 = new javax.swing.JLabel();
-        txtCodBookMod = new javax.swing.JTextField();
-        btnSearch1 = new javax.swing.JButton();
-        txtCodBookMod1 = new javax.swing.JTextField();
-        lblNewStudent17 = new javax.swing.JLabel();
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
         setUndecorated(true);
@@ -88,11 +75,11 @@ public class booksForm extends javax.swing.JFrame {
         lblStudents.setBackground(new java.awt.Color(51, 51, 51));
         lblStudents.setFont(new java.awt.Font("Liberation Serif", 1, 36)); // NOI18N
         lblStudents.setForeground(new java.awt.Color(255, 255, 255));
-        lblStudents.setText("Libros");
-        jPanel1.add(lblStudents, new org.netbeans.lib.awtextra.AbsoluteConstraints(340, 50, -1, -1));
+        lblStudents.setText("Estudiantes");
+        jPanel1.add(lblStudents, new org.netbeans.lib.awtextra.AbsoluteConstraints(300, 50, -1, -1));
 
         btnExit.setFont(new java.awt.Font("Liberation Serif", 0, 24)); // NOI18N
-        btnExit.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/goBack.png"))); // NOI18N
+        btnExit.setIcon(new javax.swing.ImageIcon(getClass().getResource("/com/mycompany/library/images/goBack.png"))); // NOI18N
         btnExit.setText("Regresar");
         btnExit.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
@@ -104,7 +91,7 @@ public class booksForm extends javax.swing.JFrame {
         jPanel2.setBackground(new java.awt.Color(0, 51, 102));
 
         btnAdd.setFont(new java.awt.Font("Liberation Serif", 0, 24)); // NOI18N
-        btnAdd.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/addBook.png"))); // NOI18N
+        btnAdd.setIcon(new javax.swing.ImageIcon(getClass().getResource("/com/mycompany/library/images/addStudent.png"))); // NOI18N
         btnAdd.setText("Agregar");
         btnAdd.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
@@ -115,22 +102,21 @@ public class booksForm extends javax.swing.JFrame {
         lblNewStudent1.setBackground(new java.awt.Color(255, 255, 255));
         lblNewStudent1.setFont(new java.awt.Font("Liberation Serif", 0, 18)); // NOI18N
         lblNewStudent1.setForeground(new java.awt.Color(255, 255, 255));
-        lblNewStudent1.setText("Código:");
+        lblNewStudent1.setText("Nombre:");
 
-        txtTitle.setFont(new java.awt.Font("Liberation Serif", 2, 18)); // NOI18N
+        txtName.setFont(new java.awt.Font("Liberation Serif", 2, 18)); // NOI18N
 
-        txtAutor.setFont(new java.awt.Font("Liberation Serif", 2, 18)); // NOI18N
+        txtCarnet.setFont(new java.awt.Font("Liberation Serif", 2, 18)); // NOI18N
 
         lblNewStudent2.setBackground(new java.awt.Color(255, 255, 255));
         lblNewStudent2.setFont(new java.awt.Font("Liberation Serif", 0, 18)); // NOI18N
         lblNewStudent2.setForeground(new java.awt.Color(255, 255, 255));
-        lblNewStudent2.setText("Autor:");
+        lblNewStudent2.setText("Carnet:");
 
         lblNewStudent3.setBackground(new java.awt.Color(255, 255, 255));
         lblNewStudent3.setFont(new java.awt.Font("Liberation Serif", 0, 18)); // NOI18N
         lblNewStudent3.setForeground(new java.awt.Color(255, 255, 255));
-        lblNewStudent3.setText("Título:");
-        lblNewStudent3.setToolTipText("");
+        lblNewStudent3.setText("Código de carrera:");
 
         txtdd.setFont(new java.awt.Font("Liberation Serif", 2, 18)); // NOI18N
 
@@ -149,12 +135,12 @@ public class booksForm extends javax.swing.JFrame {
         lblNewStudent4.setBackground(new java.awt.Color(255, 255, 255));
         lblNewStudent4.setFont(new java.awt.Font("Liberation Serif", 0, 24)); // NOI18N
         lblNewStudent4.setForeground(new java.awt.Color(255, 255, 255));
-        lblNewStudent4.setText("Nuevo libro");
+        lblNewStudent4.setText("Nuevo estudiante");
 
         lblNewStudent8.setBackground(new java.awt.Color(255, 255, 255));
         lblNewStudent8.setFont(new java.awt.Font("Liberation Serif", 0, 18)); // NOI18N
         lblNewStudent8.setForeground(new java.awt.Color(255, 255, 255));
-        lblNewStudent8.setText("Fecha de publicación:");
+        lblNewStudent8.setText("Fecha de nacimiento:");
 
         lblNewStudent9.setBackground(new java.awt.Color(255, 255, 255));
         lblNewStudent9.setFont(new java.awt.Font("Liberation Serif", 0, 18)); // NOI18N
@@ -171,37 +157,8 @@ public class booksForm extends javax.swing.JFrame {
         lblNewStudent11.setForeground(new java.awt.Color(255, 255, 255));
         lblNewStudent11.setText("año");
 
-        lblNewStudent12.setBackground(new java.awt.Color(255, 255, 255));
-        lblNewStudent12.setFont(new java.awt.Font("Liberation Serif", 0, 18)); // NOI18N
-        lblNewStudent12.setForeground(new java.awt.Color(255, 255, 255));
-        lblNewStudent12.setText("Importar desde un archivo de texto");
-
-        btnImport.setFont(new java.awt.Font("Liberation Serif", 0, 24)); // NOI18N
-        btnImport.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/importData.png"))); // NOI18N
-        btnImport.setText("Importar");
-        btnImport.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btnImportActionPerformed(evt);
-            }
-        });
-
-        txtCodBook.setFont(new java.awt.Font("Liberation Serif", 2, 18)); // NOI18N
-
-        lblNewStudent13.setBackground(new java.awt.Color(255, 255, 255));
-        lblNewStudent13.setFont(new java.awt.Font("Liberation Serif", 0, 18)); // NOI18N
-        lblNewStudent13.setForeground(new java.awt.Color(255, 255, 255));
-        lblNewStudent13.setText("No. Copias:");
-        lblNewStudent13.setToolTipText("");
-
-        txtNoCopies.setFont(new java.awt.Font("Liberation Serif", 2, 18)); // NOI18N
-
-        lblNewStudent14.setBackground(new java.awt.Color(255, 255, 255));
-        lblNewStudent14.setFont(new java.awt.Font("Liberation Serif", 0, 18)); // NOI18N
-        lblNewStudent14.setForeground(new java.awt.Color(255, 255, 255));
-        lblNewStudent14.setText("Editorial:");
-        lblNewStudent14.setToolTipText("");
-
-        txtEditorial.setFont(new java.awt.Font("Liberation Serif", 2, 18)); // NOI18N
+        jcbCodCarrera.setFont(new java.awt.Font("Liberation Serif", 0, 14)); // NOI18N
+        jcbCodCarrera.setModel(new javax.swing.DefaultComboBoxModel<>(new String[] { "1 - Ingeniería", "2 - Medicina", "3 - Derecho", "4 - Arquitectura", "5 - Administración" }));
 
         javax.swing.GroupLayout jPanel2Layout = new javax.swing.GroupLayout(jPanel2);
         jPanel2.setLayout(jPanel2Layout);
@@ -231,38 +188,25 @@ public class booksForm extends javax.swing.JFrame {
                                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                                 .addComponent(txtyyyy, javax.swing.GroupLayout.PREFERRED_SIZE, 55, javax.swing.GroupLayout.PREFERRED_SIZE))
                             .addComponent(lblNewStudent11, javax.swing.GroupLayout.Alignment.TRAILING))
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 133, Short.MAX_VALUE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 136, Short.MAX_VALUE)
                         .addComponent(btnAdd, javax.swing.GroupLayout.PREFERRED_SIZE, 210, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel2Layout.createSequentialGroup()
-                        .addComponent(lblNewStudent2)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                        .addComponent(txtAutor, javax.swing.GroupLayout.PREFERRED_SIZE, 250, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                        .addComponent(lblNewStudent14)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                        .addComponent(txtEditorial, javax.swing.GroupLayout.PREFERRED_SIZE, 200, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addGap(103, 103, 103))
                     .addGroup(jPanel2Layout.createSequentialGroup()
                         .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addComponent(lblNewStudent4)
                             .addGroup(jPanel2Layout.createSequentialGroup()
-                                .addComponent(lblNewStudent1)
+                                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                    .addComponent(lblNewStudent1)
+                                    .addComponent(lblNewStudent2))
                                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                .addComponent(txtCodBook, javax.swing.GroupLayout.PREFERRED_SIZE, 150, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                .addComponent(lblNewStudent3)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                                .addComponent(txtTitle, javax.swing.GroupLayout.PREFERRED_SIZE, 249, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                .addComponent(lblNewStudent13)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                .addComponent(txtNoCopies, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                        .addGap(0, 0, Short.MAX_VALUE))
-                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel2Layout.createSequentialGroup()
-                        .addGap(0, 0, Short.MAX_VALUE)
-                        .addComponent(lblNewStudent12)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(btnImport, javax.swing.GroupLayout.PREFERRED_SIZE, 210, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                                    .addGroup(jPanel2Layout.createSequentialGroup()
+                                        .addComponent(txtCarnet, javax.swing.GroupLayout.PREFERRED_SIZE, 150, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                                        .addComponent(lblNewStudent3)
+                                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                                        .addComponent(jcbCodCarrera, 0, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                                    .addComponent(txtName, javax.swing.GroupLayout.PREFERRED_SIZE, 462, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                            .addComponent(lblNewStudent4))
+                        .addGap(0, 0, Short.MAX_VALUE)))
                 .addContainerGap())
         );
         jPanel2Layout.setVerticalGroup(
@@ -270,49 +214,34 @@ public class booksForm extends javax.swing.JFrame {
             .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel2Layout.createSequentialGroup()
                 .addGap(24, 24, 24)
                 .addComponent(lblNewStudent4)
+                .addGap(18, 18, 18)
+                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(lblNewStudent1)
+                    .addComponent(txtName, javax.swing.GroupLayout.PREFERRED_SIZE, 28, javax.swing.GroupLayout.PREFERRED_SIZE))
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(jPanel2Layout.createSequentialGroup()
-                        .addComponent(lblNewStudent1)
-                        .addGap(19, 19, 19)
-                        .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(lblNewStudent2)
-                            .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                                .addComponent(lblNewStudent14)
-                                .addComponent(txtEditorial, javax.swing.GroupLayout.PREFERRED_SIZE, 28, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addComponent(txtAutor, javax.swing.GroupLayout.PREFERRED_SIZE, 28, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                        .addGap(5, 5, 5)
-                        .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(txtdd, javax.swing.GroupLayout.PREFERRED_SIZE, 28, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(lblNewStudent6)
-                            .addComponent(txtmm, javax.swing.GroupLayout.PREFERRED_SIZE, 28, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(lblNewStudent7)
-                            .addComponent(txtyyyy, javax.swing.GroupLayout.PREFERRED_SIZE, 28, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(lblNewStudent8))
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(lblNewStudent9)
-                            .addComponent(lblNewStudent10)
-                            .addComponent(lblNewStudent11))
-                        .addGap(0, 0, Short.MAX_VALUE))
-                    .addGroup(jPanel2Layout.createSequentialGroup()
-                        .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(txtCodBook, javax.swing.GroupLayout.PREFERRED_SIZE, 28, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(lblNewStudent3)
-                            .addComponent(txtTitle, javax.swing.GroupLayout.PREFERRED_SIZE, 28, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                                .addComponent(lblNewStudent13)
-                                .addComponent(txtNoCopies, javax.swing.GroupLayout.PREFERRED_SIZE, 28, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                        .addComponent(btnAdd)))
+                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(lblNewStudent2)
+                    .addComponent(txtCarnet, javax.swing.GroupLayout.PREFERRED_SIZE, 28, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(lblNewStudent3)
+                    .addComponent(jcbCodCarrera, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
                 .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(btnImport)
-                    .addComponent(lblNewStudent12))
-                .addContainerGap())
+                    .addComponent(txtdd, javax.swing.GroupLayout.PREFERRED_SIZE, 28, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(lblNewStudent6)
+                    .addComponent(txtmm, javax.swing.GroupLayout.PREFERRED_SIZE, 28, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(lblNewStudent7)
+                    .addComponent(txtyyyy, javax.swing.GroupLayout.PREFERRED_SIZE, 28, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(lblNewStudent8)
+                    .addComponent(btnAdd))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(lblNewStudent9)
+                    .addComponent(lblNewStudent10)
+                    .addComponent(lblNewStudent11))
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
 
-        jPanel1.add(jPanel2, new org.netbeans.lib.awtextra.AbsoluteConstraints(30, 110, 740, 250));
+        jPanel1.add(jPanel2, new org.netbeans.lib.awtextra.AbsoluteConstraints(30, 110, 740, 220));
 
         jPanel3.setBackground(new java.awt.Color(0, 51, 102));
 
@@ -328,33 +257,33 @@ public class booksForm extends javax.swing.JFrame {
         lblNewStudent5.setToolTipText("");
 
         jcbFilter.setFont(new java.awt.Font("Liberation Serif", 0, 14)); // NOI18N
-        jcbFilter.setModel(new javax.swing.DefaultComboBoxModel<>(new String[] { "Cod Libro", "Nombre", "Autor" }));
+        jcbFilter.setModel(new javax.swing.DefaultComboBoxModel<>(new String[] { "Nombre", "Carnet", "Cod Carrera" }));
 
         txtFilter.setFont(new java.awt.Font("Liberation Serif", 2, 18)); // NOI18N
 
-        tblBooks.setModel(new javax.swing.table.DefaultTableModel(
+        jTable2.setModel(new javax.swing.table.DefaultTableModel(
             new Object [][] {
-                {null, null, null, null, null, null},
-                {null, null, null, null, null, null},
-                {null, null, null, null, null, null},
-                {null, null, null, null, null, null}
+                {null, null, null, null},
+                {null, null, null, null},
+                {null, null, null, null},
+                {null, null, null, null}
             },
             new String [] {
-                "Código", "Título", "No. Copias", "Autor", "Editorial", "Fecha de publicación"
+                "Carnet", "Nombre", "Cod Carrera", "Fecha de nacimiento"
             }
         ) {
             boolean[] canEdit = new boolean [] {
-                false, false, false, false, true, true
+                false, false, false, false
             };
 
             public boolean isCellEditable(int rowIndex, int columnIndex) {
                 return canEdit [columnIndex];
             }
         });
-        tblStudents.setViewportView(tblBooks);
+        tblStudents.setViewportView(jTable2);
 
         btnSearch.setFont(new java.awt.Font("Liberation Serif", 0, 24)); // NOI18N
-        btnSearch.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/search.png"))); // NOI18N
+        btnSearch.setIcon(new javax.swing.ImageIcon(getClass().getResource("/com/mycompany/library/images/search.png"))); // NOI18N
         btnSearch.setText("Buscar");
 
         javax.swing.GroupLayout jPanel3Layout = new javax.swing.GroupLayout(jPanel3);
@@ -394,78 +323,11 @@ public class booksForm extends javax.swing.JFrame {
                         .addContainerGap()
                         .addComponent(btnSearch)))
                 .addGap(18, 18, 18)
-                .addComponent(tblStudents, javax.swing.GroupLayout.DEFAULT_SIZE, 103, Short.MAX_VALUE)
+                .addComponent(tblStudents, javax.swing.GroupLayout.DEFAULT_SIZE, 263, Short.MAX_VALUE)
                 .addContainerGap())
         );
 
-        jPanel1.add(jPanel3, new org.netbeans.lib.awtextra.AbsoluteConstraints(30, 490, 740, 220));
-
-        jPanel4.setBackground(new java.awt.Color(0, 51, 102));
-
-        lblNewStudent15.setBackground(new java.awt.Color(255, 255, 255));
-        lblNewStudent15.setFont(new java.awt.Font("Liberation Serif", 0, 24)); // NOI18N
-        lblNewStudent15.setForeground(new java.awt.Color(255, 255, 255));
-        lblNewStudent15.setText("Modificar No. Copias");
-
-        lblNewStudent16.setBackground(new java.awt.Color(255, 255, 255));
-        lblNewStudent16.setFont(new java.awt.Font("Liberation Serif", 0, 18)); // NOI18N
-        lblNewStudent16.setForeground(new java.awt.Color(255, 255, 255));
-        lblNewStudent16.setText("Código del libro:");
-        lblNewStudent16.setToolTipText("");
-
-        txtCodBookMod.setFont(new java.awt.Font("Liberation Serif", 2, 18)); // NOI18N
-
-        btnSearch1.setFont(new java.awt.Font("Liberation Serif", 0, 24)); // NOI18N
-        btnSearch1.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/modify.png"))); // NOI18N
-        btnSearch1.setText("Modificar");
-
-        txtCodBookMod1.setFont(new java.awt.Font("Liberation Serif", 2, 18)); // NOI18N
-
-        lblNewStudent17.setBackground(new java.awt.Color(255, 255, 255));
-        lblNewStudent17.setFont(new java.awt.Font("Liberation Serif", 0, 18)); // NOI18N
-        lblNewStudent17.setForeground(new java.awt.Color(255, 255, 255));
-        lblNewStudent17.setText("Nuevo No. Copias:");
-        lblNewStudent17.setToolTipText("");
-
-        javax.swing.GroupLayout jPanel4Layout = new javax.swing.GroupLayout(jPanel4);
-        jPanel4.setLayout(jPanel4Layout);
-        jPanel4Layout.setHorizontalGroup(
-            jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel4Layout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(jPanel4Layout.createSequentialGroup()
-                        .addComponent(lblNewStudent15)
-                        .addGap(0, 0, Short.MAX_VALUE))
-                    .addGroup(jPanel4Layout.createSequentialGroup()
-                        .addComponent(lblNewStudent16)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(txtCodBookMod, javax.swing.GroupLayout.PREFERRED_SIZE, 100, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                        .addComponent(lblNewStudent17)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(txtCodBookMod1, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 59, Short.MAX_VALUE)
-                        .addComponent(btnSearch1, javax.swing.GroupLayout.PREFERRED_SIZE, 210, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addContainerGap())
-        );
-        jPanel4Layout.setVerticalGroup(
-            jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel4Layout.createSequentialGroup()
-                .addGap(24, 24, 24)
-                .addComponent(lblNewStudent15)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(lblNewStudent16)
-                    .addComponent(txtCodBookMod, javax.swing.GroupLayout.PREFERRED_SIZE, 28, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(btnSearch1)
-                    .addGroup(jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                        .addComponent(lblNewStudent17)
-                        .addComponent(txtCodBookMod1, javax.swing.GroupLayout.PREFERRED_SIZE, 28, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-
-        jPanel1.add(jPanel4, new org.netbeans.lib.awtextra.AbsoluteConstraints(30, 370, 740, 110));
+        jPanel1.add(jPanel3, new org.netbeans.lib.awtextra.AbsoluteConstraints(30, 340, 740, 380));
 
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
         getContentPane().setLayout(layout);
@@ -495,10 +357,6 @@ public class booksForm extends javax.swing.JFrame {
         // TODO add your handling code here:
     }//GEN-LAST:event_btnAddActionPerformed
 
-    private void btnImportActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnImportActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_btnImportActionPerformed
-
     /**
      * @param args the command line arguments
      */
@@ -516,23 +374,20 @@ public class booksForm extends javax.swing.JFrame {
                 }
             }
         } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(booksForm.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+            java.util.logging.Logger.getLogger(StudentsForm.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(booksForm.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+            java.util.logging.Logger.getLogger(StudentsForm.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(booksForm.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+            java.util.logging.Logger.getLogger(StudentsForm.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(booksForm.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+            java.util.logging.Logger.getLogger(StudentsForm.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         }
-        //</editor-fold>
-        //</editor-fold>
-        //</editor-fold>
         //</editor-fold>
 
         /* Create and display the form */
         java.awt.EventQueue.invokeLater(new Runnable() {
             public void run() {
-                new booksForm().setVisible(true);
+                new StudentsForm().setVisible(true);
             }
         });
     }
@@ -540,24 +395,17 @@ public class booksForm extends javax.swing.JFrame {
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JButton btnAdd;
     private javax.swing.JButton btnExit;
-    private javax.swing.JButton btnImport;
     private javax.swing.JButton btnSearch;
-    private javax.swing.JButton btnSearch1;
     private javax.swing.JPanel jPanel1;
     private javax.swing.JPanel jPanel2;
     private javax.swing.JPanel jPanel3;
-    private javax.swing.JPanel jPanel4;
+    private javax.swing.JTable jTable2;
+    private javax.swing.JComboBox<String> jcbCodCarrera;
     private javax.swing.JComboBox<String> jcbFilter;
     private javax.swing.JLabel lblNewStudent;
     private javax.swing.JLabel lblNewStudent1;
     private javax.swing.JLabel lblNewStudent10;
     private javax.swing.JLabel lblNewStudent11;
-    private javax.swing.JLabel lblNewStudent12;
-    private javax.swing.JLabel lblNewStudent13;
-    private javax.swing.JLabel lblNewStudent14;
-    private javax.swing.JLabel lblNewStudent15;
-    private javax.swing.JLabel lblNewStudent16;
-    private javax.swing.JLabel lblNewStudent17;
     private javax.swing.JLabel lblNewStudent2;
     private javax.swing.JLabel lblNewStudent3;
     private javax.swing.JLabel lblNewStudent4;
@@ -567,16 +415,10 @@ public class booksForm extends javax.swing.JFrame {
     private javax.swing.JLabel lblNewStudent8;
     private javax.swing.JLabel lblNewStudent9;
     private javax.swing.JLabel lblStudents;
-    private javax.swing.JTable tblBooks;
     private javax.swing.JScrollPane tblStudents;
-    private javax.swing.JTextField txtAutor;
-    private javax.swing.JTextField txtCodBook;
-    private javax.swing.JTextField txtCodBookMod;
-    private javax.swing.JTextField txtCodBookMod1;
-    private javax.swing.JTextField txtEditorial;
+    private javax.swing.JTextField txtCarnet;
     private javax.swing.JTextField txtFilter;
-    private javax.swing.JTextField txtNoCopies;
-    private javax.swing.JTextField txtTitle;
+    private javax.swing.JTextField txtName;
     private javax.swing.JTextField txtdd;
     private javax.swing.JTextField txtmm;
     private javax.swing.JTextField txtyyyy;
